requirements.txt
# Core requirements for eBL Photo Stitcher
opencv-python>=4.5.0
numpy>=1.19.0
imageio>=2.9.0
rawpy>=0.16.0
pyexiv2>=2.8.0
cairosvg>=2.5.2
attrs>=21.4.0
typing-extensions>=4.0.0
